SD-38: Миграция на новый пайплайн
Полностью вырезал старую логику и попытался сделать текущую логику расширяемой. Получилось не так хорошо, как хотел, но в принципе пока пойдет.  Точками расширения становятся файлы model_factory и compress. В первом можно добавить новые модели(предварительно расширив энум), во втором описать их использования. В целом текущий Больше ООП подход позволяет в будущем перейти на нормальные фабрики и абстракции для классов компрессионных моделей. Так что вопрос еще остается актуальным, но начальные шаги для его решения заложены в рамках задачи
