--- #11/compress.py
@@ -1,6 +1,0 @@
-def run_coder():
-    ...
-
-
-def run_decoder():
-    ...
