+++ common/logging_sd.py
@@ -2,7 +2,8 @@ import logging
 import sys
 import os
 
-DEBUG = False
+from constants.constant import DEBUG
+
 
 def configure_logger(name):
     if DEBUG:
@@ -11,14 +12,14 @@ def configure_logger(name):
         logger.setLevel(logging.DEBUG)
 
         formatter = logging.Formatter(
-            "%(name)s | %(asctime)s | %(filename)s:%(lineno)d | %(levelname)s |"
-            " %(message)s")
+                "%(name)s | %(asctime)s | %(filename)s:%(lineno)d | %(levelname)s |"
+                " %(message)s")
 
         if not os.path.exists('data/logs/'):
             os.makedirs('data/logs/', exist_ok=True)
 
         file_handler = logging.FileHandler(
-            f"data/logs/{module_name}.log", mode='w')
+                f"data/logs/{module_name}.log", mode='w')
         file_handler.setFormatter(formatter)
         logger.addHandler(file_handler)
 
@@ -32,13 +33,13 @@ def configure_logger(name):
         logger.setLevel(logging.INFO)
 
         formatter = logging.Formatter(
-            "%(name)s | %(asctime)s | %(levelname)s | %(message)s")
+                "%(name)s | %(asctime)s | %(levelname)s | %(message)s")
 
         if not os.path.exists('data/logs/'):
             os.makedirs('data/logs/', exist_ok=True)
 
         file_handler = logging.FileHandler(
-            f"data/logs/{module_name}.log", mode='w')
+                f"data/logs/{module_name}.log", mode='w')
         file_handler.setLevel(logging.ERROR)
         file_handler.setFormatter(formatter)
         logger.addHandler(file_handler)
