--- download_ns.py
@@ -1,45 +1,0 @@
-import torch
-from diffusers import AutoencoderKL, PNDMScheduler, UNet2DConditionModel
-from transformers import CLIPTextModel, CLIPTokenizer
-
-pretrained_model_name_or_path = "stabilityai/stable-diffusion-2-1"
-torch_device = "mps"
-huggingface_token = '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'
-
-# Инициализация автоэнкодера с KL-дивергенцией из заранее обученной модели
-vae = AutoencoderKL.from_pretrained(
-        pretrained_model_name_or_path, subfolder="vae", use_auth_token=huggingface_token
-).to(torch_device)
-
-# Инициализация условной модели U-Net из заранее обученной модели
-unet = UNet2DConditionModel.from_pretrained(
-        pretrained_model_name_or_path, subfolder="unet", use_auth_token=huggingface_token
-).to(torch_device)
-
-# Инициализация планировщика Pseudo-нормального диффузионного процесса (Pseudo-Normal Diffusion Process Scheduler)
-scheduler = PNDMScheduler(
-        beta_start=0.00085, beta_end=0.012, beta_schedule="scaled_linear",
-        num_train_timesteps=1000, skip_prk_steps=True
-)
-
-# Инициализация текстовой модели CLIP из заранее обученной модели
-text_encoder = CLIPTextModel.from_pretrained(
-        pretrained_model_name_or_path, subfolder="text_encoder", use_auth_token=huggingface_token,
-)
-
-# Инициализация токенизатора CLIPTokenizer из заранее обученной модели
-tokenizer = CLIPTokenizer.from_pretrained(
-        pretrained_model_name_or_path,
-        subfolder="tokenizer",
-        use_auth_token=huggingface_token,
-        torch_dtype=torch.float16
-)
-
-# Создание тензора без условия
-uncond_input = tokenizer([""], padding="max_length", max_length=tokenizer.model_max_length, return_tensors="pt")
-# Вычисление его эмбеддинга с помощью CLIPTextModel
-with torch.no_grad():
-    uncond_embeddings = text_encoder(uncond_input.input_ids)[0].to(torch_device)
-
-# Код на Python инициализирует несколько моделей глубокого обучения из заранее
-# обученных моделей, используя Hugging Face Transformers.
