+++ metrics/metrics.py
@@ -1,18 +1,16 @@
-import numpy as np
-from sklearn.metrics.pairwise import cosine_similarity
-from skimage.metrics import structural_similarity
-from scipy.spatial.distance import hamming
 import cv2
-import math
+import numpy as np
+from scipy.spatial.distance import hamming
+from skimage.metrics import structural_similarity
+from sklearn.metrics.pairwise import cosine_similarity
 
 
 def cosine_similarity_metric(image1, image2):
-
     image1_vector = np.array(image1).ravel()
     image2_vector = np.array(image2).ravel()
 
     similarity_score = cosine_similarity(
-        [image1_vector], [image2_vector])[0][0]
+            [image1_vector], [image2_vector])[0][0]
 
     return similarity_score
 
@@ -28,11 +26,11 @@ def mse_metric(image1, image2):
 
 
 def ssim(image1, image2):
-    """Needs checking! Probably wrong!"""
+    """TODO: Needs checking! Probably wrong!"""
     image1 = cv2.cvtColor(image1, cv2.COLOR_BGR2GRAY)
     image2 = cv2.cvtColor(image2, cv2.COLOR_BGR2GRAY)
 
-    score = structural_similarity(image1, image2, data_range = image2.max() - image2.min())
+    score = structural_similarity(image1, image2, data_range=image2.max() - image2.min())
 
     return score
 
@@ -43,6 +41,6 @@ def cor_pirson(data1, data2):
     std1 = data1.std()
     std2 = data2.std()
 
-    corr = ((data1-mean1)*(data2-mean2)).mean()/(std1*std2)
+    corr = ((data1 - mean1) * (data2 - mean2)).mean() / (std1 * std2)
     # corr = ((data1 * data2).mean() - mean1 * mean2) / (std1 * std2)
     return corr
