+++ stable_diffusion/constant.py
@@ -1,0 +1,4 @@
+PRETRAINED_MODEL_NAME_OR_PATH = "stabilityai/stable-diffusion-2-1"
+TORCH_DEVICE = "mps"
+HUGGINGFACE_TOKEN = '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'
+MAXSTAPEDENOISE = 3
