+++ utils.py
@@ -1,62 +1,79 @@
 import os
+import time
 
 import cv2
 import numpy as np
-from PIL import Image
-from numpy import ndarray
 
-scaled_size_default = (1200, 1200)
-input_data_path = "data/input"
-output_data_path = "data/output"
+from constants.constant import DIR_PATH_OUTPUT, INPUT_DATA_PATH_FROM_UTILS
+from metrics import metrics
+from common.logging_sd import configure_logger
 
-
-def search_dir (path: str = input_data_path):
-    for dir_name in os.listdir(path):
-        f = os.path.join(path, dir_name)
-        if os.path.isdir(f):
-            yield f, dir_name
+logger = configure_logger(__name__)
 
 
-def load_image (path: str = input_data_path):
-    for filename in os.listdir(path):
-        f = os.path.join(path, filename)
-        if os.path.isfile(f):
-            yield f, filename
+def create_dir(new_dir_name: str, index: str = ""):
+    try:
+        os.makedirs(f"{DIR_PATH_OUTPUT}/{new_dir_name}/")
+    except FileNotFoundError:
+        logger.error(f"failed to create directory {DIR_PATH_OUTPUT}/{new_dir_name}", exc_info=True)
 
 
-def get_rescaled_img (img_path: str, scaled_size=scaled_size_default):
-    if img_path.__contains__('.png') or img_path.__contains__('.jpg'):
-        img = Image.open(img_path)
-        # img = ImageOps.fit(img, scaled_size, Image.ANTIALIAS)
-        img.thumbnail(scaled_size, Image.ANTIALIAS)
-        return img
+def load_image(path: str = INPUT_DATA_PATH_FROM_UTILS):
+    try:
+        for filename in os.listdir(path):
+            if filename.endswith('.png') or filename.endswith('.jpg'):
+                f = os.path.join(path, filename)
+                logger.debug(f"{f}, {filename}")
+                if os.path.isfile(f):
+                    yield f, filename
+        logger.debug(f"Frame search success in {path} directory")
+    except FileNotFoundError:
+        logger.error(f"Error while reading image from directory {path}, catalog not found", exc_info=True)
+
+
+def save_img(img, path: str, name_img: str = 'default'):
+    if os.path.exists(f'{DIR_PATH_OUTPUT}/{path}'):
+        try:
+            cv2.imwrite(f'{DIR_PATH_OUTPUT}/{path}/0_{name_img}', img)
+            logger.debug(f"The compressed image {name_img} was "
+                         f"saved successfully in the directory {DIR_PATH_OUTPUT}/{path}")
+        except FileNotFoundError:
+            logger.error(f"Failed to save images {name_img} to the directory {DIR_PATH_OUTPUT}/{path}, "
+                         f"file is corrupted", exc_info=True)
     else:
-        return None
+        logger.error(f"Failed to save images {name_img} to the directory {DIR_PATH_OUTPUT}/{path}, catalog not found",
+                     exc_info=True)
 
 
-def get_rescaled_img_using_cv2 (img_path: str, scaled_size=scaled_size_default):
-    if img_path.__contains__('.png') or img_path.__contains__('.jpg'):
-        image = cv2.imread(img_path)
-        res = cv2.resize(image, scaled_size)
-        return res
+def get_rescaled_cv2(image, scaled_size: tuple = (512, 512)):
+    res = cv2.resize(image, scaled_size)
+    logger.debug(f'file compression in the directory successfully')
+    return res
 
 
-def rescaled_pil_img_and_write_using_cv2 (img, path):
-    rescaled_img = cv2.resize(np.array(img), (1920, 1080))
-    cv2.imwrite(path, rescaled_img)
+def write_metrics_in_file(path: str, data: tuple, image_name: str, time: time):
+    try:
+        data_str = f"image_name: {image_name}\n" \
+                   f"ssim_data = {data[0]}\n" \
+                   f"pirson_data = {data[1]}\n" \
+                   f"cosine_similarity = {data[2]}\n" \
+                   f"mse = {data[3]}\n" \
+                   f"hamming_distance = {data[4]}\n" \
+                   f"frame_compression_time = {time}\n"
+        with open(f"{path}/metrics.txt", mode='w') as f:
+            f.write(data_str)
+    except FileNotFoundError:
+        logger.error(f"Failed to save metrics to the directory {path}, catalog not found", exc_info=True)
 
 
-def write_txt (path, data: ndarray):
-    data.tofile(path, sep=" ", format="%s")
-
-
-def write_metrics (path: str, data: str):
-    with open(path, 'w') as f:
-        f.write(data)
-
-
-# it is testing run for util function
-if __name__ == '__main__':
-    for i, ii in search_dir():
-        for j, jj in load_image(i):
-            print(j)
+def metrics_img(image, denoised_img) -> tuple:
+    img1 = (np.array(image).ravel())
+    img2 = (np.array(denoised_img).ravel())
+    ssim_data = metrics.ssim(image, denoised_img)
+    pirson_data = metrics.cor_pirson(img1, img2)
+    cosine_similarity = metrics.cosine_similarity_metric(img1, img2)
+    mse = metrics.mse_metric(image, denoised_img)
+    hamming_distance = metrics.hamming_distance_metric(image, denoised_img)
+    logger.debug(f'Collecting image metrics successfully')
+    result_metrics: tuple = (ssim_data, pirson_data, cosine_similarity, mse, hamming_distance)
+    return result_metrics
